Move sidebar for images to the image annotation folder
